+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Full Keycloak upstream jira issue can be shown if filtered by Fix version.
 - SAML/ OIDC IdP AutoUpdate
 - External introspection endpoint [EOSC-KC-140](https://github.com/eosc-kc/keycloak/issues/140)
 - Id token lifespan [RCIAM-930](https://jira.argo.grnet.gr/browse/RCIAM-930)
+- Autoupdated SAML Client [RCIAM-1181](https://jira.argo.grnet.gr/browse/RCIAM-1181)
 
 ### Changed
 - Change emailVerified User field with UserAttributeMappers (conditional trust email). [EOSC-KC-70](https://github.com/eosc-kc/keycloak/issues/70)